Fixes for JAMA-based SVD.
Fixes Issue 16

M    sspace/src/edu/ucla/sspace/common/matrix/DiagonalMatrix.java

- set() now throws an error if the user attempts to set a non-diagonal element.
  (Previously it just ignored this)

M    sspace/src/edu/ucla/sspace/common/MatrixIO.java

- Added DENSE_TEXT support to writeMatrix

M    sspace/src/edu/ucla/sspace/common/SVD.java

- Rewrote parts of jamaSVD() to manually truncate the returned matrices to the
  correct number of dimensions.

- Once the matrices are returned, we no longer have an additional write-to-file
  step, which was unnecessary in the first place

M    sspace/src/edu/ucla/sspace/common/Matrices.java

- Added support for creating a resized version of a Matrix (currently unused)

- Fixed formatting inconsistencies

M    sspace/src/edu/ucla/sspace/ri/RandomIndexing.java

- Reverted back to autoboxing, as the bytecode confirms the calls to valueOf()
  and the autoboxed version looks much cleaner.  This is a purely syntactic
  change.

+++ sspace/src/edu/ucla/sspace/ri/RandomIndexing.java
@@ -748,16 +748,12 @@ public class RandomIndexing implements SemanticSpace {
 
 	    for (int p : v.positiveDimensions()) {
 		Integer count = sparseArray.get(p);
-		sparseArray.put(p, (count == null) 
-				? Integer.valueOf(1) 
-				: Integer.valueOf(count.intValue() + 1));
+		sparseArray.put(p, (count == null) ? 1 : count + 1);
 	    }
 		
 	    for (int n : v.negativeDimensions()) {
 		Integer count = sparseArray.get(n);
-		sparseArray.put(n, (count == null) 
-				? Integer.valueOf(1) 
-				: Integer.valueOf(count.intValue() + 1));		
+		sparseArray.put(n, (count == null) ? 1 : count + 1);		
 	    }		
 	}
